+++ src/layout/index.tsx
@@ -1,5 +1,6 @@
 import { TrackedScrollUpButton } from "@/browser/scroll-up";
-import type { PropsWithChildren } from "react";
+import type { NextPage } from "next";
+import type { PropsWithChildren, ReactNode } from "react";
 import { Footer } from "./footer";
 import type { HeaderProps } from "./header";
 import { Header } from "./header";
@@ -20,3 +21,14 @@ export function Layout({ children, header }: PropsWithChildren<LayoutProps>) {
     </div>
   );
 }
+
+type GetLayout = ({ children }: PropsWithChildren) => ReactNode;
+export type NextPageWithLayout<Props = unknown, InitialProps = Props> = NextPage<
+  Props,
+  InitialProps
+> & {
+  noAuth?: boolean;
+  getLayout?: GetLayout;
+};
+
+export const defaultGetLayout: GetLayout = ({ children }) => children;
