+++ src/pages/404.tsx
@@ -1,5 +1,5 @@
 import { Error } from "@/components/error";
-import type { NextPageWithLayout } from "@/utils/type";
+import type { NextPageWithLayout } from "@/layout";
 
 const Page: NextPageWithLayout = function () {
   return <Error message="Page not found" />;
