+++ src/pages/_error.tsx
@@ -1,5 +1,5 @@
 import { Error } from "@/components/error";
-import type { NextPageWithLayout } from "@/utils/type";
+import type { NextPageWithLayout } from "@/layout";
 
 const Page: NextPageWithLayout = function () {
   return <Error message="Something went wrong" />;
