+++ src/pages/auth/error.tsx
@@ -1,6 +1,6 @@
 import { getQueryParam } from "@/browser/query";
 import { Error } from "@/components/error";
-import type { NextPageWithLayout } from "@/utils/type";
+import type { NextPageWithLayout } from "@/layout";
 import { useRouter } from "next/router";
 
 const Page: NextPageWithLayout = function () {
